spacetimedb_bindings_sys

Function console_log

pub fn console_log(
    level: LogLevel,
    target: Option<&str>,
    filename: Option<&str>,
    line_number: Option<u32>,
    text: &str,
)

Log at level a text message occurring in filename:line_number with target being the module path at the log! invocation site.