spacetimedb_bindings_sys

Function datastore_delete_by_index_scan_point_bsatn

pub fn datastore_delete_by_index_scan_point_bsatn(
    index_id: IndexId,
    point: &[u8],
) -> Result<u32>

Deletes all rows found in the index identified by index_id, according to the `point.

This syscall will delete all the rows found by datastore_index_scan_point_bsatn with the same arguments passed. See datastore_index_scan_point_bsatn for details.

The number of rows deleted is returned on success.

Errors

Returns an error:

NOT_IN_TRANSACTION, when called outside of a transaction.

NO_SUCH_INDEX, when index_id is not a known ID of an index.

WRONG_INDEX_ALGO if the index is not a range-compatible index.

BSATN_DECODE_ERROR, when point cannot be decoded to an AlgebraicValue typed at the index’s key type (AlgebraicType).