spacetimedb_bindings_sys

Function datastore_delete_by_index_scan_range_bsatn

pub fn datastore_delete_by_index_scan_range_bsatn(
    index_id: IndexId,
    prefix: &[u8],
    prefix_elems: ColId,
    rstart: &[u8],
    rend: &[u8],
) -> Result<u32, Errno>

Deletes all rows found in the index identified by index_id, according to the prefix, rstart, and rend.

This syscall will delete all the rows found by datastore_index_scan_range_bsatn with the same arguments passed, including prefix_elems. See datastore_index_scan_range_bsatn for details.

The number of rows deleted is returned on success.

Errors

Returns an error:

NOT_IN_TRANSACTION, when called outside of a transaction.

NO_SUCH_INDEX, when index_id is not a known ID of an index.

WRONG_INDEX_ALGO if the index is not a range-compatible index.

BSATN_DECODE_ERROR, when prefix cannot be decoded to a prefix_elems number of AlgebraicValue typed at the initial prefix_elems AlgebraicTypes of the index’s key type. Or when rstart or rend cannot be decoded to an Bound<AlgebraicValue> where the inner AlgebraicValues are typed at the prefix_elems + 1 AlgebraicType of the index’s key type.