spacetimedb_bindings_sys

Function datastore_update_bsatn

pub fn datastore_update_bsatn(
    table_id: TableId,
    index_id: IndexId,
    row: &mut [u8],
) -> Result<&[u8], Errno>

Updates a row into the table identified by table_id, where the row is a BSATN-encoded ProductValue matching the table’s ProductType row-schema.

The row to update is found by projecting row to the type of the unique index identified by index_id. If no row is found, row is inserted.

The row is &mut due to auto-incrementing columns. So row is written to with the updated row re-encoded.

Returns an error if

a table with the provided table_id doesn’t exist

an index with the provided index_id doesn’t exist or if the index was not unique.

there were unique constraint violations

row doesn’t decode from BSATN to a ProductValue according to the ProductType that the table’s schema specifies or if row cannot project to the index’s type.

the row was not found