spacetimedb_bindings_sys

Module raw

Provides a raw set of sys calls which abstractions can be built atop of.

Structs

BytesSink

A handle into a buffer of bytes in the host environment that can be written to.

BytesSource

A handle into a buffer of bytes in the host environment that can be read from.

RowIter

Represents table iterators.

Enums

IndexType

What strategy does the database index use?

Constants

LOG_LEVEL_DEBUG

The debug log level. See console_log.

LOG_LEVEL_ERROR

The error log level. See console_log.

LOG_LEVEL_INFO

The info log level. See console_log.

LOG_LEVEL_PANIC

The panic log level. See console_log.

LOG_LEVEL_TRACE

The trace log level. See console_log.

LOG_LEVEL_WARN

The warn log level. See console_log.

Functions

bytes_sink_write⚠

Writes up to buffer_len bytes from buffer = buffer_ptr[..buffer_len], to the sink, registered in the host environment.

bytes_source_read⚠

Reads bytes from source, registered in the host environment, and stores them in the memory pointed to by buffer = buffer_ptr[..buffer_len].

console_log⚠

Logs at level a message message occurring in filename:line_number with target being the module path at the log! invocation site.

console_timer_end⚠

End a timing span.

console_timer_start⚠

Begins a timing span with name = name_ptr[..name_len].

datastore_btree_scan_bsatn⚠Deprecated

This is the same as datastore_index_scan_range_bsatn.

datastore_delete_all_by_eq_bsatn⚠

Deletes those rows, in the table identified by table_id, that match any row in the byte string rel = rel_ptr[..rel_len] in WASM memory.

datastore_delete_by_btree_scan_bsatn⚠Deprecated

This is the same as datastore_delete_by_index_scan_range_bsatn.

datastore_delete_by_index_scan_range_bsatn⚠

Deletes all rows found in the index identified by index_id, according to the:

datastore_index_scan_range_bsatn⚠

Finds all rows in the index identified by index_id, according to the:

datastore_insert_bsatn⚠

Inserts a row into the table identified by table_id, where the row is read from the byte string row = row_ptr[..row_len] in WASM memory where row_len = row_len_ptr[..size_of::<usize>()] stores the capacity of row.

datastore_table_row_count⚠

Writes the number of rows currently in table identified by table_id to out.

datastore_table_scan_bsatn⚠

Starts iteration on each row, as BSATN-encoded, of a table identified by table_id.

datastore_update_bsatn⚠

Updates a row in the table identified by table_id to row where the row is read from the byte string row = row_ptr[..row_len] in WASM memory where row_len = row_len_ptr[..size_of::<usize>()] stores the capacity of row.

identity⚠

Writes the identity of the module into out = out_ptr[..32].

index_id_from_name⚠

Queries the index_id associated with the given (index) name where name is the UTF-8 slice in WASM memory at name_ptr[..name_len].

row_iter_bsatn_advance⚠

Reads rows from the given iterator registered under iter.

row_iter_bsatn_close⚠

Destroys the iterator registered under iter.

table_id_from_name⚠

Queries the table_id associated with the given (table) name where name is the UTF-8 slice in WASM memory at name_ptr[..name_len].