spacetimedb_bindings_sys::raw

Function bytes_source_read

pub unsafe extern "C" fn bytes_source_read(
    source: BytesSource,
    buffer_ptr: *mut u8,
    buffer_len_ptr: *mut usize,
) -> i16

Reads bytes from source, registered in the host environment, and stores them in the memory pointed to by buffer = buffer_ptr[..buffer_len].

The buffer_len = buffer_len_ptr[..size_of::<usize>()] stores the capacity of buffer. On success (0 or -1 is returned), buffer_len is set to the number of bytes written to buffer. When -1 is returned, the resource has been exhausted and there are no more bytes to read, leading to the resource being immediately destroyed. Note that the host is free to reuse allocations in a pool, destroying the handle logically does not entail that memory is necessarily reclaimed.

Traps

Traps if:

buffer_len_ptr is NULL or buffer_len is not in bounds of WASM memory.

buffer_ptr is NULL or buffer is not in bounds of WASM memory.

Errors

Returns an error:

NO_SUCH_BYTES, when source is not a valid bytes source.

Example

The typical use case for this ABI is in __call_reducer__, to read and deserialize the args. An example definition, dealing with args might be:

/// #[no_mangle]
extern "C" fn __call_reducer__(..., args: BytesSource, ...) -> i16 {
    // ...

    let mut buf = Vec::<u8>::with_capacity(1024);
    loop {
        // Write into the spare capacity of the buffer.
        let buf_ptr = buf.spare_capacity_mut();
        let spare_len = buf_ptr.len();
        let mut buf_len = buf_ptr.len();
        let buf_ptr = buf_ptr.as_mut_ptr().cast();
        let ret = unsafe { bytes_source_read(args, buf_ptr, &mut buf_len) };
        // SAFETY: `bytes_source_read` just appended `spare_len` bytes to `buf`.
        unsafe { buf.set_len(buf.len() + spare_len) };
        match ret {
            // Host side source exhausted, we're done.
            -1 => break,
            // Wrote the entire spare capacity.
            // Need to reserve more space in the buffer.
            0 if spare_len == buf_len => buf.reserve(1024),
            // Host didn't write as much as possible.
            // Try to read some more.
            // The host will likely not trigger this branch,
            // but a module should be prepared for it.
            0 => {}
            _ => unreachable!(),
        }
    }

    // ...
}