spacetimedb_bindings_sys::raw

Function console_log

pub unsafe extern "C" fn console_log(
    level: u8,
    target_ptr: *const u8,
    target_len: usize,
    filename_ptr: *const u8,
    filename_len: usize,
    line_number: u32,
    message_ptr: *const u8,
    message_len: usize,
)

Logs at level a message message occurring in filename:line_number with target being the module path at the log! invocation site.

These various pointers are interpreted lossily as UTF-8 strings with a corresponding _len.

The target and filename pointers are ignored by passing NULL. The line number is ignored if line_number == u32::MAX.

No message is logged if

target != NULL && target + target_len > u64::MAX

filename != NULL && filename + filename_len > u64::MAX

message + message_len > u64::MAX

Traps

Traps if:

target is not NULL and target_ptr[..target_len] is not in bounds of WASM memory.

filename is not NULL and filename_ptr[..filename_len] is not in bounds of WASM memory.

message is not NULL and message_ptr[..message_len] is not in bounds of WASM memory.