spacetimedb_bindings_sys::raw

Function console_timer_end

pub unsafe extern "C" fn console_timer_end(timer_id: u32) -> u16

End a timing span.

The timer_id must be the result of a call to console_timer_start. The duration between the two calls will be computed and printed to the module’s logs. Once console_timer_end is called on id: ConsoleTimerId, the id is invalid. That is, console_timer_end(id) the second time will yield NO_SUCH_CONSOLE_TIMER.

Note that the host is free to reuse allocations in a pool, destroying the handle logically does not entail that memory is necessarily reclaimed.

Errors

Returns an error:

NO_SUCH_CONSOLE_TIMER, when timer_id does not exist.