spacetimedb_bindings_sys::raw

Function console_timer_start

pub unsafe extern "C" fn console_timer_start(
    name_ptr: *const u8,
    name_len: usize,
) -> u32

Begins a timing span with name = name_ptr[..name_len].

When the returned ConsoleTimerId is passed to console_timer_end, the duration between the calls will be printed to the module’s logs.

The name is interpreted lossily as UTF-8.

Traps

Traps if:

name_ptr is NULL or name is not in bounds of WASM memory.