spacetimedb_bindings_sys::raw

Function datastore_delete_all_by_eq_bsatn

pub unsafe extern "C" fn datastore_delete_all_by_eq_bsatn(
    table_id: TableId,
    rel_ptr: *const u8,
    rel_len: usize,
    out: *mut u32,
) -> u16

Deletes those rows, in the table identified by table_id, that match any row in the byte string rel = rel_ptr[..rel_len] in WASM memory.

Matching is defined by first BSATN-decoding the byte string pointed to at by relation to a Vec<ProductValue> according to the row schema of the table and then using Ord for AlgebraicValue. A match happens when Ordering::Equal is returned from fn cmp. This occurs exactly when the row’s BSATN-encoding is equal to the encoding of the ProductValue.

The number of rows deleted is written to the WASM pointer out.

Traps

Traps if:

rel_ptr is NULL or rel is not in bounds of WASM memory.

out is NULL or out[..size_of::<u32>()] is not in bounds of WASM memory.

Errors

Returns an error:

NOT_IN_TRANSACTION, when called outside of a transaction.

NO_SUCH_TABLE, when table_id is not a known ID of a table.

BSATN_DECODE_ERROR, when rel cannot be decoded to Vec<ProductValue> where each ProductValue is typed at the ProductType the table’s schema specifies.