spacetimedb_bindings_sys::raw

Function datastore_index_scan_range_bsatn

pub unsafe extern "C" fn datastore_index_scan_range_bsatn(
    index_id: IndexId,
    prefix_ptr: *const u8,
    prefix_len: usize,
    prefix_elems: ColId,
    rstart_ptr: *const u8,
    rstart_len: usize,
    rend_ptr: *const u8,
    rend_len: usize,
    out: *mut RowIter,
) -> u16

Finds all rows in the index identified by index_id, according to the:

prefix = prefix_ptr[..prefix_len],

rstart = rstart_ptr[..rstart_len],

rend = rend_ptr[..rend_len],

in WASM memory.

The index itself has a schema/type. The prefix is decoded to the initial prefix_elems AlgebraicTypes whereas rstart and rend are decoded to the prefix_elems + 1 AlgebraicType where the AlgebraicValues are wrapped in Bound. That is, rstart, rend are BSATN-encoded Bound<AlgebraicValue>s.

Matching is then defined by equating prefix to the initial prefix_elems columns of the index and then imposing rstart as the starting bound and rend as the ending bound on the prefix_elems + 1 column of the index. Remaining columns of the index are then unbounded. Note that the prefix in this case can be empty (prefix_elems = 0), in which case this becomes a ranged index scan on a single-col index or even a full table scan if rstart and rend are both unbounded.

The relevant table for the index is found implicitly via the index_id, which is unique for the module.

On success, the iterator handle is written to the out pointer. This handle can be advanced by row_iter_bsatn_advance.

Non-obvious queries

For an index on columns [a, b, c]:

a = x, b = y is encoded as a prefix [x, y] and a range Range::Unbounded, or as a prefix [x] and a range rstart = rend = Range::Inclusive(y).

a = x, b = y, c = z is encoded as a prefix [x, y] and a range rstart = rend = Range::Inclusive(z).

A sorted full scan is encoded as an empty prefix and a range Range::Unbounded.

Traps

Traps if:

prefix_elems > 0 and (prefix_ptr is NULL or prefix is not in bounds of WASM memory).

rstart is NULL or rstart is not in bounds of WASM memory.

rend is NULL or rend is not in bounds of WASM memory.

out is NULL or out[..size_of::<RowIter>()] is not in bounds of WASM memory.

Errors

Returns an error:

NOT_IN_TRANSACTION, when called outside of a transaction.

NO_SUCH_INDEX, when index_id is not a known ID of an index.

WRONG_INDEX_ALGO if the index is not a range-compatible index.

BSATN_DECODE_ERROR, when prefix cannot be decoded to a prefix_elems number of AlgebraicValue typed at the initial prefix_elems AlgebraicTypes of the index’s key type. Or when rstart or rend cannot be decoded to an Bound<AlgebraicValue> where the inner AlgebraicValues are typed at the prefix_elems + 1 AlgebraicType of the index’s key type.