spacetimedb_bindings_sys::raw

Function datastore_update_bsatn

pub unsafe extern "C" fn datastore_update_bsatn(
    table_id: TableId,
    index_id: IndexId,
    row_ptr: *mut u8,
    row_len_ptr: *mut usize,
) -> u16

Updates a row in the table identified by table_id to row where the row is read from the byte string row = row_ptr[..row_len] in WASM memory where row_len = row_len_ptr[..size_of::<usize>()] stores the capacity of row.

The byte string row must be a BSATN-encoded ProductValue typed at the table’s ProductType row-schema.

The row to update is found by projecting row to the type of the unique index identified by index_id. If no row is found, the error NO_SUCH_ROW is returned.

To handle auto-incrementing columns, when the call is successful, the row is written back to with the generated sequence values. These values are written as a BSATN-encoded pv: ProductValue. Each v: AlgebraicValue in pv is typed at the sequence’s column type. The vs in pv are ordered by the order of the columns, in the schema of the table. When the table has no sequences, this implies that the pv, and thus row, will be empty. The row_len is set to the length of bsatn(pv).

Traps

Traps if:

row_len_ptr is NULL or row_len is not in bounds of WASM memory.

row_ptr is NULL or row is not in bounds of WASM memory.

Errors

Returns an error:

NOT_IN_TRANSACTION, when called outside of a transaction.

NO_SUCH_TABLE, when table_id is not a known ID of a table.

NO_SUCH_INDEX, when index_id is not a known ID of an index.

INDEX_NOT_UNIQUE, when the index was not unique.

NO_SUCH_ROW, when the row was not found in the unique index.

BSATN_DECODE_ERROR, when row cannot be decoded to a ProductValue typed at the ProductType the table’s schema specifies or when it cannot be projected to the index identified by index_id.

UNIQUE_ALREADY_EXISTS, when inserting row would violate a unique constraint.

SCHEDULE_AT_DELAY_TOO_LONG, when the delay specified in the row was too long.