spacetimedb_bindings_sys::raw

Function index_id_from_name

pub unsafe extern "C" fn index_id_from_name(
    name_ptr: *const u8,
    name_len: usize,
    out: *mut IndexId,
) -> u16

Queries the index_id associated with the given (index) name where name is the UTF-8 slice in WASM memory at name_ptr[..name_len].

If a name was provided for the RawIndexDef for this index, that is the name of the index. If no name was provided, a name was autogenerated like so:

let table_name: String = "my_table".into();
let column_names: Vec<String> = vec!["bananas".into(), "oranges".into()];
let column_names = column_names.join("_");
let name = format!("{table_name}_{column_names}_idx_btree");

(See the function spacetimedb_schema::def::validate::v9::generate_index_name for more information.)

The index id is written into the out pointer.

Traps

Traps if:

name_ptr is NULL or name is not in bounds of WASM memory.

name is not valid UTF-8.

out is NULL or out[..size_of::<IndexId>()] is not in bounds of WASM memory.

Errors

Returns an error:

NOT_IN_TRANSACTION, when called outside of a transaction.

NO_SUCH_INDEX, when name is not the name of an index.