spacetimedb_bindings_sys::raw

Function row_iter_bsatn_advance

pub unsafe extern "C" fn row_iter_bsatn_advance(
    iter: RowIter,
    buffer_ptr: *mut u8,
    buffer_len_ptr: *mut usize,
) -> i16

Reads rows from the given iterator registered under iter.

Takes rows from the iterator and stores them in the memory pointed to by buffer = buffer_ptr[..buffer_len], encoded in BSATN format.

The buffer_len = buffer_len_ptr[..size_of::<usize>()] stores the capacity of buffer. On success (0 or -1 is returned), buffer_len is set to the combined length of the encoded rows. When -1 is returned, the iterator has been exhausted and there are no more rows to read, leading to the iterator being immediately destroyed. Note that the host is free to reuse allocations in a pool, destroying the handle logically does not entail that memory is necessarily reclaimed.

Traps

Traps if:

buffer_len_ptr is NULL or buffer_len is not in bounds of WASM memory.

buffer_ptr is NULL or buffer is not in bounds of WASM memory.

Errors

Returns an error:

NO_SUCH_ITER, when iter is not a valid iterator.

BUFFER_TOO_SMALL, when there are rows left but they cannot fit in buffer. When this occurs, buffer_len is set to the size of the next item in the iterator. To make progress, the caller should reallocate the buffer to at least that size and try again.